==> 01c_Globaltest.Rmd <==
---
title: "Globaltest"
author: "John Zhuang"
date: "2024-06-04"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```



```{r}
library(globaltest)
library(ggplot2)
library(dplyr)
```

# 0. Data Preparation

```{r}
data.horizontal = readRDS("data/derived/data_horizontal_joined.rds")
```

Keep only columns relevant for globaltest (all numerical datasets) and convert them into matrices:

```{r}
gt.matrix = as.matrix(data.horizontal[,1:701])
```

# 1. Globaltest

```{r}
Y = data.horizontal$vital_status
```
```{r, fig.width=70, fig.height=10}
x.mRSG = gt.matrix[, 1:470]
gt.mRSG = covariates(gt(Y ~ x.mRSG, model = "logistic"),
                     cluster = F, help.lines = T, alias = colnames(x.mRSG),
                     cex.labels = 0.5, colors = c(2,3))
```
```{r, fig.width=20, fig.height=10}
x.R2Gn = gt.matrix[, 471:668]
gt.R2Gn = covariates(gt(Y ~ x.R2Gn, model = "logistic"),
                     cluster = F, help.lines = T, alias = colnames(x.R2Gn),
                     cex.labels = 0.5, colors = c(2,3))
```
```{r, fig.width=15, fig.height=10}
x.RPPA = gt.matrix[, 669:701]
gt.RPPA = covariates(gt(Y ~ x.RPPA, model = "logistic"),
                     cluster = F, help.lines = T, alias = colnames(x.RPPA),
                     cex.labels = 0.7, colors = c(2,3))
```

## A. Zoomed-in miRNASeqGene and RNASeq2GeneNorm and compare all selections with selected features in glmnet and squeezy

Since the plots for miRNASeqGene and RNASeq2GeneNorm contain too many variables,
we will zoom in on the ones with the smallest p values.

Both plots, as well as the full plot for RPPAArray features, will be compared with features selected in Glmnet and Squeezy.

```{r}
# Selects all overlapping features from
merge.all.sel.features = function(featurelist){
  out = c()
  for (i in 1:length(featurelist)) {
    for (j in 1:length(featurelist[[i]])) {
      for (k in 1:length(featurelist[[i]][[j]])) {
    out = c(out, names(featurelist[[i]][[j]][[k]]))
      }
    }
  }
  out.u = unique(out)
  out.u
}
```

```{r}
logistic.glmnet.features = merge.all.sel.features(readRDS("data/sel_features/results_glmnet_coef.rds"))
cox.glmnet.features = merge.all.sel.features(readRDS("data/sel_features/results_glmnet_cox_coef_1se.rds"))
squeezy.features = merge.all.sel.features(readRDS("data/sel_features/results_squeezy_coef.rds"))
```

```{r}
print(logistic.glmnet.features)
sum(startsWith(logistic.glmnet.features, "hsa")) #mRSG Counts
sum(endsWith(logistic.glmnet.features, "R2Gn")) #R2Gn Counts
sum(endsWith(logistic.glmnet.features, "RPPA")) #RPPA Counts
```

```{r}
print(cox.glmnet.features)
sum(startsWith(cox.glmnet.features, "hsa")) #mRSG Counts
sum(endsWith(cox.glmnet.features, "R2Gn")) #R2Gn Counts
sum(endsWith(cox.glmnet.features, "RPPA")) #RPPA Counts
```
```{r}
print(squeezy.features)
sum(startsWith(squeezy.features, "hsa")) #mRSG Counts
sum(endsWith(squeezy.features, "R2Gn")) #R2Gn Counts
sum(endsWith(squeezy.features, "RPPA")) #RPPA Counts
```



```{r}
# Return the top most significant results from a Globaltest covariate object
# Select all significant p values with a limit of 0.05 (default), or any other limit
# or select the top n most significant variables (overwrites limit if not NA)
get.significant.pvalues = function(gtcov, limit = 0.05, n = NA) {
  res = result(gtcov)[,c("alias", "p-value")]
  dir = ((extract(gtcov))@extra)[,c("alias", "direction")]
  out = right_join(res, dir, by = "alias")
  names(out) = c("var", "p.value", "direction")
  out = out[order(out$p.value),] # Order by p-value
  #out$p.1.minus = 1 - out$p.value # Use 1-p for plotting
  if (!is.na(n)) {
    out = out[1:n,]  # Select the first (value) variables with the smallest p-values
  } else {
    out = out[out$p.value <= limit,] # Cut off based on p-value limit
  }
  out
}
```


```{r}
# Add features selected from penalized regression to selected gtcov variables
gtp.pr = function(gtcov, limit = 0.05, n = NA) {
  gtp = get.significant.pvalues(gtcov, limit, n)
  gtp$logistic.glmnet = as.numeric(gtp$var %in% logistic.glmnet.features) * 0.7
  gtp$squeezy = as.numeric(gtp$var %in% squeezy.features) * 0.8
  gtp$cox.glmnet = as.numeric(gtp$var %in% cox.glmnet.features) * 0.9
  gtp$logistic.glmnet[gtp$logistic.glmnet == 0] = NA
  gtp$squeezy[gtp$squeezy == 0] = NA
  gtp$cox.glmnet[gtp$cox.glmnet == 0] = NA
  gtp
}
```

```{r}
# Manually visualize the dataframe selected from get.significant.pvalues
# Optional: horizontal line drawing at selected p value.
gtcov.manual = function(gtp, limit = NULL) {
  p = ggplot(data = gtp, aes(x=reorder(var, p.value), y=log10(p.value), fill = direction)) +
    geom_bar(stat="identity", show.legend = T) + 
    geom_text(aes(y=log10(logistic.glmnet)), label = "L", fontface = "bold") +
    geom_text(aes(y=log10(squeezy)), label = "S", fontface = "bold") +
    geom_text(aes(y=log10(cox.glmnet)), label = "C", fontface = "bold") +
    scale_y_reverse(labels = c(0.001, 0.01, 0.1, 1)) +
    scale_x_discrete(guide = guide_axis(angle = 90)) +
    #theme(axis.text.x=element_text(size=rel(0.5), angle=90)) +
    scale_fill_manual(values = c(3,2)) +
    #scale_y_log10()
    xlab("") +
    ylab("p-value") +
    theme_minimal()
  
  if (!is.null(limit)) {
    p = p + geom_hline(yintercept=log10(limit), linetype="dashed", 
                color = "black")
  }
  p
}
```





For miRNASeqGene, we plot all variables that correspond to a significant p value (<0.05).

For RNASeq2GeneNorm, we plot the top 70 smallest p values (not all of them are significant).

The threashold alpha = 0.05 is plotted as a dashed black line.

```{r, fig.width=15, fig.height=10}
gtcov.manual(gtp.pr(gt.mRSG), limit = 0.05)
gtcov.manual(gtp.pr(gt.R2Gn, n = 70), limit = 0.05)
```

```{r, fig.width=15, fig.height=10}
p.value = limit = 0.05
p = ggplot(data = gtp.pr(gt.RPPA, limit = 1), aes(x=reorder(var, p.value), y=log10(p.value), fill = direction)) +
    geom_bar(stat="identity", show.legend = T) + 
    geom_text(aes(y=log10(logistic.glmnet)), label = "L", fontface = "bold") +
    geom_text(aes(y=log10(squeezy)), label = "S", fontface = "bold") +
    geom_text(aes(y=log10(cox.glmnet)), label = "C", fontface = "bold") +
    scale_y_reverse(labels = c(0.003, 0.01, 0.03, 0.1, 0.3, 1)) +
    #scale_y_reverse() +
    scale_x_discrete(guide = guide_axis(angle = 90)) +
    #theme(axis.text.x=element_text(size=rel(0.5), angle=90)) +
    scale_fill_manual(values = c(3,2)) +
    #scale_y_log10()
    xlab("") +
    ylab("p-value") +
    theme_minimal()
  
  if (!is.null(limit)) {
    p = p + geom_hline(yintercept=log10(limit), linetype="dashed", 
                color = "black")
  }
  p
```

### Report number of selected features per dataset

```{r}
table(gtp.pr(gt.mRSG)$logistic.glmnet)
table(gtp.pr(gt.mRSG)$cox.glmnet)
table(gtp.pr(gt.mRSG)$squeezy)

table(gtp.pr(gt.R2Gn)$logistic.glmnet)
table(gtp.pr(gt.R2Gn)$cox.glmnet)
table(gtp.pr(gt.R2Gn)$squeezy)

table(gtp.pr(gt.RPPA)$logistic.glmnet)
table(gtp.pr(gt.RPPA)$cox.glmnet)
table(gtp.pr(gt.RPPA)$squeezy)
```



## B. Save significant (P <= 0.05) to external file

```{r}
results.globaltest = list("miRNASeqGene" = get.significant.pvalues(gt.mRSG),
                          "RNASeq2GeneNorm" = get.significant.pvalues(gt.R2Gn),
                          "RPPAArray" = get.significant.pvalues(gt.RPPA))
saveRDS(results.globaltest, file = "data/sel_features/results_globaltest_sig.rds")
```

## C. Save Globaltest models to external file

```{r}
models.globaltest = list("miRNASeqGene" = gt.mRSG,
                          "RNASeq2GeneNorm" = gt.R2Gn,
                          "RPPAArray" = gt.RPPA)
saveRDS(models.globaltest, file = "data/out_models/results_globaltest.rds")
```
