==> 02c_MOFA.Rmd <==
---
title: "MOFA"
author: "John Zhuang"
date: "2024-05-16"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(MOFA2)
library(utils)
library(stats)
library(ggplot2)
library(dplyr) # for case_when
```

```{r}
data.vertical = readRDS("data/derived/data_vertical_numeric_only.rds")
```

# 1. MOFA Training

## A. Check again to reassure if colnames are identical before analysis
```{r}
#
identical(colnames(data.vertical$RPPAArray), colnames(data.vertical$RNASeq2GeneNorm))
identical(colnames(data.vertical$RNASeq2GeneNorm), colnames(data.vertical$miRNASeqGene))
```
## B. Create MOFA object

```{r}
matrix.list = list("view_mRNA" = as.matrix(data.vertical$miRNASeqGene),
                   "view_R2Gn" = as.matrix(data.vertical$RNASeq2GeneNorm),
                   "view_RPPA" = as.matrix(data.vertical$RPPAArray))
```

```{r}
data.mofa = create_mofa(matrix.list)
```
```{r}
plot_data_overview(data.mofa)
```

## C. Add metadata to the model

```{r}
data.horizontal = readRDS("data/derived/data_horizontal.rds")
meta.data = data.horizontal$colData.selected
```

```{r}
#Rename patientID to sample for MOFA
names(meta.data)[1] = "sample"

samples_metadata(data.mofa) = meta.data
head(samples_metadata(data.mofa))
```

## D. Model training parameters
```{r}
data.opts = get_default_data_options(data.mofa)
data.opts
model.opts = get_default_model_options(data.mofa)
model.opts
training.opts = get_default_training_options(data.mofa)
training.opts
```
The output warns the default 15 factors to be too many for the sample size.
Thus, we reduce it to the 10 (maximum recommended):

```{r}
model.opts_10 = get_default_model_options(data.mofa)
model.opts_10$num_factors = 10
```

We will also do 1, 2, 3 and 5 factors in order to compare results with those FABIA.

```{r}
model.opts_1 = get_default_model_options(data.mofa)
model.opts_2 = get_default_model_options(data.mofa)
model.opts_3 = get_default_model_options(data.mofa)
model.opts_5 = get_default_model_options(data.mofa)

model.opts_1$num_factors = 1
model.opts_2$num_factors = 2
model.opts_3$num_factors = 3
model.opts_5$num_factors = 5
```

## E. Train the model.

### 10 factors

```{r}
data.mofa.ready_10 = prepare_mofa(
  object = data.mofa,
  data_options = data.opts,
  model_options = model.opts_10,
  training_options = training.opts
)
```
```{r, cache=F}
outfile = "data/out_models/mofa10f.hdf5"
set.seed(42)
MOFA.10f <- run_mofa(data.mofa.ready_10, outfile, use_basilisk=T)
```


### 1 factor

```{r}
data.mofa.ready_1 = prepare_mofa(
  object = data.mofa,
  data_options = data.opts,
  model_options = model.opts_1,
  training_options = training.opts
)
```

```{r, cache=F}
outfile = "data/out_models/mofa1f.hdf5"
set.seed(42)
MOFA.1f <- run_mofa(data.mofa.ready_1, outfile, use_basilisk=T)
```


### 2 factors

```{r}
data.mofa.ready_2 = prepare_mofa(
  object = data.mofa,
  data_options = data.opts,
  model_options = model.opts_2,
  training_options = training.opts
)
```

```{r, cache=F}
outfile = "data/out_models/mofa2f.hdf5"
set.seed(42)
MOFA.2f <- run_mofa(data.mofa.ready_2, outfile, use_basilisk=T)
```


### 3 factors

```{r}
data.mofa.ready_3 = prepare_mofa(
  object = data.mofa,
  data_options = data.opts,
  model_options = model.opts_3,
  training_options = training.opts
)
```

```{r, cache=F}
outfile = "data/out_models/mofa3f.hdf5"
set.seed(42)
MOFA.3f <- run_mofa(data.mofa.ready_3, outfile, use_basilisk=T)
```

### 5 factors

```{r}
data.mofa.ready_5 = prepare_mofa(
  object = data.mofa,
  data_options = data.opts,
  model_options = model.opts_5,
  training_options = training.opts
)
```

```{r, cache=F}
outfile = "data/out_models/mofa5f.hdf5"
set.seed(42)
MOFA.5f <- run_mofa(data.mofa.ready_5, outfile, use_basilisk=T)
```



```{r, eval=F}
# Loads pre-existing data for 10-fold MOFA.
# Used only for debugging and not in the rendering of final results.
filepath <- "data/out_models/mofa10f.hdf5"
MOFA.10f <- load_model(filepath)
```

```{r}
plot_data_overview(MOFA.10f)
```


# MOFA

## 1. Variance Decomposition

```{r}
get_variance_explained(MOFA.10f)
```
```{r}
plot_variance_explained(MOFA.10f, x="view", y="factor")
```

```{r}
plot_variance_explained(MOFA.10f, x="view", y="factor", plot_total = T)[[2]]
```

Plots other # of factors:

```{r}
plot_variance_explained(MOFA.1f, x="view", y="factor", plot_total = T)
plot_variance_explained(MOFA.2f, x="view", y="factor", plot_total = T)
plot_variance_explained(MOFA.3f, x="view", y="factor", plot_total = T)
plot_variance_explained(MOFA.5f, x="view", y="factor", plot_total = T)

get_variance_explained(MOFA.1f)
get_variance_explained(MOFA.2f)
get_variance_explained(MOFA.3f)
get_variance_explained(MOFA.5f)
```


## 2. Visualization of Factors

```{r}
plot_factor(MOFA.10f,
            factor = 1:10,
            color_by = "event_time",
            shape_by = "vital_status")
```

```{r}
plot_factor(MOFA.10f,
                factors = c(1:10),
                color_by = "vital_status",
                dodge = T,
                add_violin = T,
                violin_alpha = 0.1) +
  scale_fill_manual(values = c("blue", "red"))
```
```{r fig.width=10, fig.height=10}
plot_factors(MOFA.10f,
            factors = 1:10,
            color_by = "vital_status") +
  scale_fill_manual(values = c("blue", "red")) +
  scale_color_manual(values = c("blue", "red"))
```
```{r, eval = F}
plot_data_heatmap(MOFA.10f,
                  factor = 1,
                  features = 100)
```
```{r fig.width=10, fig.height=10}
plot_data_scatter(MOFA.10f,
                  factor = 1,
                  features = 20,
                  add_lm = T,
                  color_by = "vital_status") +
  scale_fill_manual(values = c("blue", "red"))
```

For other # of factors MOFA:

```{r}
plot_factor(MOFA.1f,
            factor = 1,
            color_by = "event_time",
            shape_by = "vital_status")

plot_factor(MOFA.2f,
            factor = 1:2,
            color_by = "event_time",
            shape_by = "vital_status")

plot_factor(MOFA.3f,
            factor = 1:2,
            color_by = "event_time",
            shape_by = "vital_status")

plot_factor(MOFA.5f,
            factor = 1:5,
            color_by = "event_time",
            shape_by = "vital_status")
```

```{r}
plot_factor(MOFA.1f,
                factors = 1,
                color_by = "vital_status",
                dodge = T,
                add_violin = T,
                violin_alpha = 0.1) +
  scale_fill_manual(values = c("blue", "red"))


plot_factor(MOFA.2f,
                factors = c(1:2),
                color_by = "vital_status",
                dodge = T,
                add_violin = T,
                violin_alpha = 0.1) +
  scale_fill_manual(values = c("blue", "red"))

plot_factor(MOFA.3f,
                factors = c(1:3),
                color_by = "vital_status",
                dodge = T,
                add_violin = T,
                violin_alpha = 0.1) +
  scale_fill_manual(values = c("blue", "red"))

plot_factor(MOFA.5f,
                factors = c(1:5),
                color_by = "vital_status",
                dodge = T,
                add_violin = T,
                violin_alpha = 0.1) +
  scale_fill_manual(values = c("blue", "red"))
```


```{r}
plot_factors(MOFA.1f,
            factors = 1,
            color_by = "vital_status") +
  scale_fill_manual(values = c("blue", "red")) +
  scale_color_manual(values = c("blue", "red"))

plot_factors(MOFA.2f,
            factors = 1:2,
            color_by = "vital_status") +
  scale_fill_manual(values = c("blue", "red")) +
  scale_color_manual(values = c("blue", "red"))

plot_factors(MOFA.3f,
            factors = 1:3,
            color_by = "vital_status") +
  scale_fill_manual(values = c("blue", "red")) +
  scale_color_manual(values = c("blue", "red"))

plot_factors(MOFA.5f,
            factors = 1:5,
            color_by = "vital_status") +
  scale_fill_manual(values = c("blue", "red")) +
  scale_color_manual(values = c("blue", "red"))
```

## 3. Extract factors and Weights

```{r fig.width=40, fig.height=8}
plot_weights(MOFA.10f,
             factor = 1:10,
             nfeatures = 10,
             abs = F)

plot_top_weights(MOFA.10f,
                 factor = 1:10,
                 nfeatures = 20)
```

```{r, eval=F}
factors = get_factors(MOFA.10f, as.data.frame = T)

weights = get_weights(MOFA.10f, as.data.frame = T)
```

### A. Obtain top 5%, 10% and 25% weights

```{r}
top_x_percent = function(MOFA.result, factors = NA, perct = 5, print =  T, show.weights = F) {
  quant = 1 - perct * 0.01
  MOFA.weight = get_weights(MOFA.result, as.data.frame = T)
  
  #automatically read factors from MOFA result
  
  if(is.na(factors)) {
    factors = length(table(MOFA.weight$factor))
  }
  
  out.list = vector("list", length = factors)
  names(out.list) = names(table(MOFA.weight$factor))
  
  for (i in 1:factors) {
    curr_factor = names(out.list)[i]
    w_curr = MOFA.weight[MOFA.weight$factor == curr_factor, ] #Factor1 ~ Factor10
    qt = quantile(abs(w_curr$value), quant)
    selected = w_curr[abs(w_curr$value) >= qt, c(1,3)]
    
    sel.names = selected$feature
    sel.val =  selected$value
    sel = setNames(sel.val, sel.names)
    out.list[[i]] = sel
    
    if (print) {
      print(curr_factor)
      if (show.weights) {
        print(sel)
      } else
        print(names(sel))
    }
  }
  out.list
}
```

#### Top 5%

```{r}
m10_5 = top_x_percent(MOFA.10f, perct = 5)
m1_5 = top_x_percent(MOFA.1f, perct = 5)
m2_5 = top_x_percent(MOFA.2f, perct = 5)
m3_5 = top_x_percent(MOFA.3f, perct = 5)
m5_5 = top_x_percent(MOFA.5f, perct = 5)
```

```{r}
top_5_perc_result = list("1f" = m1_5,
                         "2f" = m2_5,
                         "3f" = m3_5,
                         "5f" = m5_5,
                         "10f" = m10_5)

saveRDS(top_5_perc_result, file = "data/sel_features/results_mofa_top_5_percent.rds")
```

#### Top 10%

```{r}
m10_10 = top_x_percent(MOFA.10f, perct = 10)
m1_10 = top_x_percent(MOFA.1f, perct = 10)
m2_10 = top_x_percent(MOFA.2f, perct = 10)
m3_10 = top_x_percent(MOFA.3f, perct = 10)
m5_10 = top_x_percent(MOFA.5f, perct = 10)
```

```{r}
top_10_perc_result = list("1f" = m1_10,
                         "2f" = m2_10,
                         "3f" = m3_10,
                         "5f" = m5_10,
                         "10f" = m10_10)

saveRDS(top_10_perc_result, file = "data/sel_features/results_mofa_top_10_percent.rds")
```

#### Top 25%
```{r}
m10_25 = top_x_percent(MOFA.10f, perct = 25, print = F)
m1_25 = top_x_percent(MOFA.1f, perct = 25, print = F)
m2_25 = top_x_percent(MOFA.2f, perct = 25, print = F)
m3_25 = top_x_percent(MOFA.3f, perct = 25, print = F)
m5_25 = top_x_percent(MOFA.5f, perct = 25, print = F)
```

```{r}
top_25_perc_result = list("1f" = m1_25,
                         "2f" = m2_25,
                         "3f" = m3_25,
                         "5f" = m5_25,
                         "10f" = m10_25)

saveRDS(top_25_perc_result, file = "data/sel_features/results_mofa_top_25_percent.rds")
```

### B. Plot the top x% of the variables

Check max and min value of loadings in order to create symmetrical plots:

```{r}
max_mW = max(max(get_weights(MOFA.1f, as.data.frame = T)$value),
             max(get_weights(MOFA.2f, as.data.frame = T)$value),
             max(get_weights(MOFA.3f, as.data.frame = T)$value),
             max(get_weights(MOFA.5f, as.data.frame = T)$value),
             max(get_weights(MOFA.10f, as.data.frame = T)$value)
)
min_mW = min(min(get_weights(MOFA.1f, as.data.frame = T)$value),
             min(get_weights(MOFA.2f, as.data.frame = T)$value),
             min(get_weights(MOFA.3f, as.data.frame = T)$value),
             min(get_weights(MOFA.5f, as.data.frame = T)$value),
             min(get_weights(MOFA.10f, as.data.frame = T)$value)
)
c(max_mW, min_mW)
```

The limits of y axis will be (-1.2, 1.2) in order to make them symmetrical.

```{r}
plot.mofa.topx = function(mofa_result, mofa_top5, mofa_top10, mofa_top25, f, abs.ymax = 1.2) {
  MOFA.weight = get_weights(mofa_result, as.data.frame = T)[ ,1:3]
  MOFA.weight$feature = as.character(MOFA.weight$feature)
  
  #automatically read factors from MOFA result
  total_fs = length(table(MOFA.weight$factor))
  if (f > total_fs) {
    stop("Requested factor number exceeds the total available factors.")
  }
  Factorx = paste("Factor", f, sep = "")
  MOFA.w = MOFA.weight[MOFA.weight$factor == Factorx, ]
  
  MOFA.w$top5 = MOFA.w$feature %in% names(mofa_top5[[Factorx]])
  MOFA.w$top10 = MOFA.w$feature %in% names(mofa_top10[[Factorx]])
  MOFA.w$top25 = MOFA.w$feature %in% names(mofa_top25[[Factorx]])
  MOFA.w$top = NA
  for(i in 1:nrow(MOFA.w)) {
    if (MOFA.w[i,]$top25 == F) {
      MOFA.w[i,]$top = NA
    } else if (MOFA.w[i,]$top10 == F) {
      MOFA.w[i,]$top = "Top 11% ~ 25%"
    } else if (MOFA.w[i,]$top5 == F) {
      MOFA.w[i,]$top = "Top 06% ~ 10%"
    } else {
      MOFA.w[i,]$top = "Top 05%" 
    }
  }
  MOFA.w$dataset = sapply(MOFA.w$feature, retrieve.dataset.type)
  #MOFA.w
  ggplot(data = MOFA.w, aes(x= feature, y = value, color = top, shape=dataset)) +
  geom_point() +
  #geom_text(nudge_y = .04, alpha = 1, size = 3) +
  theme(
    axis.title.x = element_blank(),
    axis.text.x = element_blank()
  ) +
  ggtitle(paste(total_fs , "Factors")) +
  ylab(Factorx) +
  ylim(-abs.ymax, abs.ymax)
}

retrieve.dataset.type = function(var.name) {
  if (startsWith(var.name, "hsa")) {
    out = "miRNASeqGene"
  } else if (endsWith(var.name, "RPPA")) {
    out = "RPPAArray"
  } else if (endsWith(var.name, "R2Gn")) {
    out = "RNASeq2GeneNorm"
  } else {
    out = NA
  }
  out
}
```


#### 1 Factor
```{r}
plot.mofa.topx(MOFA.1f, m1_5, m1_10, m1_25, 1)
```

#### 2 Factors

```{r}
plot.mofa.topx(MOFA.2f, m2_5, m2_10, m2_25, 1)
plot.mofa.topx(MOFA.2f, m2_5, m2_10, m2_25, 2)
```

#### 3 Factors

```{r}
plot.mofa.topx(MOFA.3f, m3_5, m3_10, m3_25, 1)
plot.mofa.topx(MOFA.3f, m3_5, m3_10, m3_25, 2)
plot.mofa.topx(MOFA.3f, m3_5, m3_10, m3_25, 3)
```

#### 5 Factors

```{r}
plot.mofa.topx(MOFA.5f, m5_5, m5_10, m5_25, 1)
plot.mofa.topx(MOFA.5f, m5_5, m5_10, m5_25, 2)
plot.mofa.topx(MOFA.5f, m5_5, m5_10, m5_25, 3)
plot.mofa.topx(MOFA.5f, m5_5, m5_10, m5_25, 4)
plot.mofa.topx(MOFA.5f, m5_5, m5_10, m5_25, 5)
```

#### 10 Factors
```{r}
plot.mofa.topx(MOFA.10f, m10_5, m10_10, m10_25, 1)
plot.mofa.topx(MOFA.10f, m10_5, m10_10, m10_25, 2)
plot.mofa.topx(MOFA.10f, m10_5, m10_10, m10_25, 3)
plot.mofa.topx(MOFA.10f, m10_5, m10_10, m10_25, 4)
plot.mofa.topx(MOFA.10f, m10_5, m10_10, m10_25, 5)
plot.mofa.topx(MOFA.10f, m10_5, m10_10, m10_25, 6)
plot.mofa.topx(MOFA.10f, m10_5, m10_10, m10_25, 7)
plot.mofa.topx(MOFA.10f, m10_5, m10_10, m10_25, 8)
plot.mofa.topx(MOFA.10f, m10_5, m10_10, m10_25, 9)
plot.mofa.topx(MOFA.10f, m10_5, m10_10, m10_25, 10)
```

### C. Select features using the same calculation as in extractBic() for FABIA

Find the variable threshold from the default sample threshold of 0.5.

Backported from the FABIA-MOFA ensemble methods.

```{r}
get.MOFA.tL = function(MOFA.result, tZ.MOFA = 0.5) {
  
  
  
#Extract MOFA outputs to that of the same format as FABIA
Z.MOFA = t(MOFA.result@expectations$Z$group1)
L.MOFA.list = MOFA.result@expectations$W
L.MOFA = rbind(L.MOFA.list$view_RPPA, L.MOFA.list$view_R2Gn, L.MOFA.list$view_mRNA)



threasZ.MOFA = tZ.MOFA

tt = 0
for (i in 1:ncol(L.MOFA)) {
  tt = tt + sum(L.MOFA[,i]^2) * sum(Z.MOFA[i,]^2)
}
tt = tt/(as.double(nrow(L.MOFA)) * as.double(ncol(L.MOFA)) * as.double(ncol(Z.MOFA)))
threasL.MOFA = sqrt(tt) / threasZ.MOFA
threasL.MOFA
}
```

```{r}
select.by.Bic = function(MOFA.result, factors = NA, perct = 5, print =  T, show.threasL = T, show.weights = F) {
  
  thresL = get.MOFA.tL(MOFA.result)
  if (show.threasL) {
        print(paste("thresL: ",thresL))
  }
  
  MOFA.weight = get_weights(MOFA.result, as.data.frame = T)
  
  #automatically read factors from MOFA result
  
  if(is.na(factors)) {
    factors = length(table(MOFA.weight$factor))
  }
  
  out.list = vector("list", length = factors)
  names(out.list) = names(table(MOFA.weight$factor))
  
  for (i in 1:factors) {
    curr_factor = names(out.list)[i]
    w_curr = MOFA.weight[MOFA.weight$factor == curr_factor, ] #Factor1 ~ Factor10
    
    selected = w_curr[abs(w_curr$value) >= thresL, c(1,3)]
    
    sel.names = selected$feature
    sel.val =  selected$value
    sel = setNames(sel.val, sel.names)
    out.list[[i]] = sel
    
    if (print) {
      print(curr_factor)
      if (show.weights) {
        print(sel)
      } else
        print(names(sel))
    }
  }
  out.list
}
```

```{r}
m10_bic = select.by.Bic(MOFA.10f)
m1_bic = select.by.Bic(MOFA.1f)
m2_bic = select.by.Bic(MOFA.2f)
m3_bic = select.by.Bic(MOFA.3f)
m5_bic = select.by.Bic(MOFA.5f)
```
```{r}
extract_by_Bic_result = list("1f" = m1_bic,
                         "2f" = m2_bic,
                         "3f" = m3_bic,
                         "5f" = m5_bic,
                         "10f" = m10_bic)

saveRDS(extract_by_Bic_result, file = "data/sel_features/results_mofa_extracted_by_bic.rds")
```

#### Plotting:

```{r}
plot.mofa.selbyBic = function(mofa_result, mofa_bic, f, abs.ymax = 1.2) {
  MOFA.weight = get_weights(mofa_result, as.data.frame = T)[ ,1:3]
  MOFA.weight$feature = as.character(MOFA.weight$feature)
  
  #automatically read factors from MOFA result
  total_fs = length(table(MOFA.weight$factor))
  if (f > total_fs) {
    stop("Requested factor number exceeds the total available factors.")
  }
  Factorx = paste("Factor", f, sep = "")
  MOFA.w = MOFA.weight[MOFA.weight$factor == Factorx, ]
  
  MOFA.w$in.bic = MOFA.w$feature %in% names(mofa_bic[[Factorx]])

  MOFA.w$dataset = sapply(MOFA.w$feature, retrieve.dataset.type)
  #MOFA.w
  ggplot(data = MOFA.w, aes(x= feature, y = value, color = in.bic, shape=dataset)) +
  geom_point() +
  #geom_text(nudge_y = .04, alpha = 1, size = 3) +
  theme(
    axis.title.x = element_blank(),
    axis.text.x = element_blank()
  ) +
  ggtitle(paste(total_fs , "Factors")) +
  ylab(Factorx) +
  ylim(-abs.ymax, abs.ymax)
}

```


#### 1 Factor
```{r}
plot.mofa.selbyBic(MOFA.1f, m1_bic, 1)
```

#### 2 Factors

```{r}
plot.mofa.selbyBic(MOFA.2f, m2_bic, 1)
plot.mofa.selbyBic(MOFA.2f, m2_bic, 2)
```

#### 3 Factors

```{r}
plot.mofa.selbyBic(MOFA.3f, m3_bic, 1)
plot.mofa.selbyBic(MOFA.3f, m3_bic, 2)
plot.mofa.selbyBic(MOFA.3f, m3_bic, 3)
```

#### 5 Factors

```{r}
plot.mofa.selbyBic(MOFA.5f, m5_bic, 1)
plot.mofa.selbyBic(MOFA.5f, m5_bic, 2)
plot.mofa.selbyBic(MOFA.5f, m5_bic, 3)
plot.mofa.selbyBic(MOFA.5f, m5_bic, 4)
plot.mofa.selbyBic(MOFA.5f, m5_bic, 5)
```

#### 10 Factors
```{r}
plot.mofa.selbyBic(MOFA.10f, m10_bic, 1)
plot.mofa.selbyBic(MOFA.10f, m10_bic, 2)
plot.mofa.selbyBic(MOFA.10f, m10_bic, 3)
plot.mofa.selbyBic(MOFA.10f, m10_bic, 4)
plot.mofa.selbyBic(MOFA.10f, m10_bic, 5)
plot.mofa.selbyBic(MOFA.10f, m10_bic, 6)
plot.mofa.selbyBic(MOFA.10f, m10_bic, 7)
plot.mofa.selbyBic(MOFA.10f, m10_bic, 8)
plot.mofa.selbyBic(MOFA.10f, m10_bic, 9)
plot.mofa.selbyBic(MOFA.10f, m10_bic, 10)
```

### C. Manual Display of L threshold for all MOFAs

```{r}
print(paste(1, "Factors, thresL =", get.MOFA.tL(MOFA.1f)))
print(paste(2, "Factors, thresL =", get.MOFA.tL(MOFA.2f)))
print(paste(3, "Factors, thresL =", get.MOFA.tL(MOFA.3f)))
print(paste(5, "Factors, thresL =", get.MOFA.tL(MOFA.5f)))
print(paste(10, "Factors, thresL =", get.MOFA.tL(MOFA.10f)))
```
