==> 05h_Pathway_Analysis_Ensemble_FABIA.Rmd <==
---
title: "Pathway Analysis for FABIA Ensembles"
author: "John Zhuang"
date: "2024-06-17"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(multiMiR)
library(clusterProfiler)
library(org.Hs.eg.db)
library(stringr)
library(ggplot2)
```

# 0. Loading of Results

## A. List of all Gene names in the dataset

```{r}
gene.names = readRDS("data/derived/unique_gene_names.rds")
```

## B. Selected Features

```{r}
results.superbiclust.fabia20 = readRDS("data/sel_features/results_superbiclust.fabia.rds")
sel.features = results.superbiclust.fabia20$resNames
sel.features
```

# 1. Convert microRNA to genes

Also remove the ending ".R2Gn" and ".RPPA" from gene names.


```{r}
split.by.omic = function(variable.list) {
  out.mRSG = variable.list[startsWith(variable.list, "hsa")]
  out.R2Gn = variable.list[endsWith(variable.list, "R2Gn")]
  out.RPPA = variable.list[endsWith(variable.list, "RPPA")]
  out = list("miRNASeqGene" = out.mRSG,
             "RNASeq2GeneNorm" = out.R2Gn,
             "RPPAArray" = out.RPPA)
  out
}
```

```{r}
multiomics.to.GENES = function(sel.features, full.search = F, table = "validated", drop.miRNA = F) {
  multiomics.list = split.by.omic(sel.features)
  # Convert miRNASeqGene to target GENE names
  # Skipped and returns an empty list instead of the list of select features,
  # or if microRNA features are explicitly dropped (drop.miRNA = T).
  # (Usually done when the features has only a few miRNA and unable to target any genes)
  if (length(multiomics.list[[1]]) > 0 & !drop.miRNA) {
    if (!full.search) { # Search only among the 198 genes in ACC, recommended
    targets= get_multimir(mirna = multiomics.list[[1]], 
                          target = gene.names,
                          table = table,
                          summary = T)
    mRSG.target.genes = unique(targets@data$target_symbol)
    } else { # Search the full dataset, not recommended. 
             # Use only if ACC gene search returns nothing or fails.
             # Genes not in the ACC genes need to be manually removed.
    targets= get_multimir(mirna = multiomics.list[[1]], 
                          table = table,
                          summary = T)  
    mRSG.target.genes = unique(targets@data$target_symbol)
    }
    
  } else {
    mRSG.target.genes = list()
  }
  #Remove suffix for Pathway Analysis Compatibility
  R2Gn.genes = str_remove(multiomics.list[[2]], ".R2Gn")
  
  RPPA.genes = str_remove(multiomics.list[[3]], ".RPPA")
  
  #Join all cleaned-up genes are remove duplicates.
  out1 = c(mRSG.target.genes, R2Gn.genes, RPPA.genes)
  out = as.character(unique(out1))
  out
}
```


```{r}
geneNames <- list("biclust1"=multiomics.to.GENES(sel.features[[1]]),
                  "biclust2"=multiomics.to.GENES(sel.features[[2]]),
                  "biclust3"=multiomics.to.GENES(sel.features[[3]]),
                  "biclust4"=multiomics.to.GENES(sel.features[[4]]),
                  "biclust5"=multiomics.to.GENES(sel.features[[5]]),
                  #"biclust6"=multiomics.to.GENES(sel.features[[6]], table = "all"), # Only works with table = "all", removed
                  #"biclust7"=multiomics.to.GENES(sel.features[[7]], full.search = T), #No results found, error
                  "biclust8"=multiomics.to.GENES(sel.features[[8]]),
                  #"biclust9"=multiomics.to.GENES(sel.features[[9]], full.search = T), #No results found, error
                  "biclust10"=multiomics.to.GENES(sel.features[[10]]),
                  "biclust12"=multiomics.to.GENES(sel.features[[12]]),
                  #"biclust13"=multiomics.to.GENES(sel.features[[13]], full.search = T), #No results found, error
                  #"biclust14"=multiomics.to.GENES(sel.features[[14]], full.search = T), #No results found, error
                  "biclust16"=multiomics.to.GENES(sel.features[[16]]),
                  "biclust18"=multiomics.to.GENES(sel.features[[18]]),
                  "biclust22"=multiomics.to.GENES(sel.features[[22]]),
                  "biclust26"=multiomics.to.GENES(sel.features[[26]]),
                  "biclust28"=multiomics.to.GENES(sel.features[[28]]))
```
```{r}
geneNames
```


```{r}
saveRDS(geneNames, file = "data/sel_features/gene_converted/genes_superbiclust.fabia.rds")
```



# 2. Pathway Analysis

## A. GO Classification

```{r}
ggo_MF <- list()#functional profile of a geneset for Molecular Function
ggo_BP <- list()#functional profile of a geneset for Biological Process
ggo_CC <- list()#functional profile of a geneset for Cellular Component


for(z in 1:length(geneNames)){
  print(z)
  input <- geneNames[[z]]

  ggo_MF[[z]] <- groupGO(gene=input,
    OrgDb = org.Hs.eg.db,
    ont="MF",
    keyType="SYMBOL",
    readable=TRUE)
  
  ggo_BP[[z]] <- groupGO(gene=input,
    OrgDb = org.Hs.eg.db,
    ont="BP",
    keyType="SYMBOL",
    readable=TRUE)

  ggo_CC[[z]] <- groupGO(gene=input,
    OrgDb = org.Hs.eg.db,
    ont="CC",
    keyType="SYMBOL",
    readable=TRUE)    
}
```

```{r}
for(z in 1:length(geneNames)) {
  print(head(ggo_MF[[z]]))
  print(head(ggo_BP[[z]]))
  print(head(ggo_CC[[z]]))
}
```
```{r}
barplot(ggo_MF[[1]], title = "graphGO, Molecular Function, bicluster 1")
barplot(ggo_MF[[2]], title = "graphGO, Molecular Function, bicluster 2")
barplot(ggo_MF[[3]], title = "graphGO, Molecular Function, bicluster 3")
barplot(ggo_MF[[4]], title = "graphGO, Molecular Function, bicluster 4")
barplot(ggo_MF[[5]], title = "graphGO, Molecular Function, bicluster 5")
barplot(ggo_MF[[6]], title = "graphGO, Molecular Function, bicluster 8")
barplot(ggo_MF[[7]], title = "graphGO, Molecular Function, bicluster 10")
barplot(ggo_MF[[8]], title = "graphGO, Molecular Function, bicluster 12")
barplot(ggo_MF[[9]], title = "graphGO, Molecular Function, bicluster 16")
barplot(ggo_MF[[10]], title = "graphGO, Molecular Function, bicluster 18")
barplot(ggo_MF[[11]], title = "graphGO, Molecular Function, bicluster 22")
barplot(ggo_MF[[12]], title = "graphGO, Molecular Function, bicluster 26")
barplot(ggo_MF[[13]], title = "graphGO, Molecular Function, bicluster 28")

barplot(ggo_BP[[1]], title = "graphGO, Biological Process, bicluster 1")
barplot(ggo_BP[[2]], title = "graphGO, Biological Process, bicluster 2")
barplot(ggo_BP[[3]], title = "graphGO, Biological Process, bicluster 3")
barplot(ggo_BP[[4]], title = "graphGO, Biological Process, bicluster 4")
barplot(ggo_BP[[5]], title = "graphGO, Biological Process, bicluster 5")
barplot(ggo_BP[[6]], title = "graphGO, Biological Process, bicluster 8")
barplot(ggo_BP[[7]], title = "graphGO, Biological Process, bicluster 10")
barplot(ggo_BP[[8]], title = "graphGO, Biological Process, bicluster 12")
barplot(ggo_BP[[9]], title = "graphGO, Biological Process, bicluster 16")
barplot(ggo_BP[[10]], title = "graphGO, Biological Process, bicluster 18")
barplot(ggo_BP[[11]], title = "graphGO, Biological Process, bicluster 22")
barplot(ggo_BP[[12]], title = "graphGO, Biological Process, bicluster 26")
barplot(ggo_BP[[13]], title = "graphGO, Biological Process, bicluster 28")

barplot(ggo_CC[[1]], title = "graphGO, Cellular Component, bicluster 1")
barplot(ggo_CC[[2]], title = "graphGO, Cellular Component, bicluster 2")
barplot(ggo_CC[[3]], title = "graphGO, Cellular Component, bicluster 3")
barplot(ggo_CC[[4]], title = "graphGO, Cellular Component, bicluster 4")
barplot(ggo_CC[[5]], title = "graphGO, Cellular Component, bicluster 5")
barplot(ggo_CC[[6]], title = "graphGO, Cellular Component, bicluster 8")
barplot(ggo_CC[[7]], title = "graphGO, Cellular Component, bicluster 10")
barplot(ggo_CC[[8]], title = "graphGO, Cellular Component, bicluster 12")
barplot(ggo_CC[[9]], title = "graphGO, Cellular Component, bicluster 16")
barplot(ggo_CC[[10]], title = "graphGO, Cellular Component, bicluster 18")
barplot(ggo_CC[[11]], title = "graphGO, Cellular Component, bicluster 22")
barplot(ggo_CC[[12]], title = "graphGO, Cellular Component, bicluster 26")
barplot(ggo_CC[[13]], title = "graphGO, Cellular Component, bicluster 28")

```

## B. GO Over-Representation

```{r}
ego_MF <- list()#GO enrichment of a geneset for Molecular Function
ego_BP <- list()#GO enrichment of a geneset for Molecular Function
ego_CC <- list()#GO enrichment of a geneset for Molecular Function


for(z in 1:length(geneNames)){
  print(z)
  input <- geneNames[[z]]
  ego_MF[[z]] <- enrichGO(
    gene          = input,
    OrgDb         = org.Hs.eg.db,
    ont           = "MF",
    pAdjustMethod = "BH",
    keyType="SYMBOL",
    pvalueCutoff  = 0.05,
    qvalueCutoff  = 0.05,
    minGSSize     = 2,
    readable      = TRUE)
  
  ego_BP[[z]] <- enrichGO(
    gene          = input,
    OrgDb         = org.Hs.eg.db,
    ont           = "BP",
    pAdjustMethod = "BH",
    keyType="SYMBOL",
    pvalueCutoff  = 0.05,
    qvalueCutoff  = 0.05,
    minGSSize     = 2,
    readable      = TRUE)
  
  ego_CC[[z]] <- enrichGO(
    gene          = input,
    OrgDb         = org.Hs.eg.db,
    ont           = "CC",
    pAdjustMethod = "BH",
    keyType="SYMBOL",
    pvalueCutoff  = 0.05,
    qvalueCutoff  = 0.05,
    minGSSize     = 2,
    readable      = TRUE)
}
```

```{r}
for(z in 1:length(geneNames)){
  print(head(ego_MF[[z]]))
  print(head(ego_BP[[z]]))
  print(head(ego_CC[[z]]))
}
```


```{r}
barplot(ego_MF[[1]], title = "enrichGO, Molecular Function, bicluster 1")
barplot(ego_MF[[2]], title = "enrichGO, Molecular Function, bicluster 2")
barplot(ego_MF[[3]], title = "enrichGO, Molecular Function, bicluster 3")
barplot(ego_MF[[4]], title = "enrichGO, Molecular Function, bicluster 4")
barplot(ego_MF[[5]], title = "enrichGO, Molecular Function, bicluster 5")
barplot(ego_MF[[6]], title = "enrichGO, Molecular Function, bicluster 8")
barplot(ego_MF[[7]], title = "enrichGO, Molecular Function, bicluster 10")
barplot(ego_MF[[8]], title = "enrichGO, Molecular Function, bicluster 12")
barplot(ego_MF[[9]], title = "enrichGO, Molecular Function, bicluster 16")
barplot(ego_MF[[10]], title = "enrichGO, Molecular Function, bicluster 18")
barplot(ego_MF[[11]], title = "enrichGO, Molecular Function, bicluster 22")
barplot(ego_MF[[12]], title = "enrichGO, Molecular Function, bicluster 26")
barplot(ego_MF[[13]], title = "enrichGO, Molecular Function, bicluster 28")

barplot(ego_BP[[1]], title = "enrichGO, Biological Process, bicluster 1")
barplot(ego_BP[[2]], title = "enrichGO, Biological Process, bicluster 2")
barplot(ego_BP[[3]], title = "enrichGO, Biological Process, bicluster 3")
barplot(ego_BP[[4]], title = "enrichGO, Biological Process, bicluster 4")
barplot(ego_BP[[5]], title = "enrichGO, Biological Process, bicluster 5")
barplot(ego_BP[[6]], title = "enrichGO, Biological Process, bicluster 8")
barplot(ego_BP[[7]], title = "enrichGO, Biological Process, bicluster 10")
barplot(ego_BP[[8]], title = "enrichGO, Biological Process, bicluster 12")
barplot(ego_BP[[9]], title = "enrichGO, Biological Process, bicluster 16")
barplot(ego_BP[[10]], title = "enrichGO, Biological Process, bicluster 18")
barplot(ego_BP[[11]], title = "enrichGO, Biological Process, bicluster 22")
barplot(ego_BP[[12]], title = "enrichGO, Biological Process, bicluster 26")
barplot(ego_BP[[13]], title = "enrichGO, Biological Process, bicluster 28")

barplot(ego_CC[[1]], title = "enrichGO, Cellular Component, bicluster 1")
barplot(ego_CC[[2]], title = "enrichGO, Cellular Component, bicluster 2")
barplot(ego_CC[[3]], title = "enrichGO, Cellular Component, bicluster 3")
barplot(ego_CC[[4]], title = "enrichGO, Cellular Component, bicluster 4")
barplot(ego_CC[[5]], title = "enrichGO, Cellular Component, bicluster 5")
barplot(ego_CC[[6]], title = "enrichGO, Cellular Component, bicluster 8")
barplot(ego_CC[[7]], title = "enrichGO, Cellular Component, bicluster 10")
barplot(ego_CC[[8]], title = "enrichGO, Cellular Component, bicluster 12")
barplot(ego_CC[[9]], title = "enrichGO, Cellular Component, bicluster 16")
barplot(ego_CC[[10]], title = "enrichGO, Cellular Component, bicluster 18")
barplot(ego_CC[[11]], title = "enrichGO, Cellular Component, bicluster 22")
barplot(ego_CC[[12]], title = "enrichGO, Cellular Component, bicluster 26")
barplot(ego_CC[[13]], title = "enrichGO, Cellular Component, bicluster 28")

```

### DAG plots for ego

```{r fig.width=20, fig.height=10}
goplot(ego_MF[[1]]) + labs(title = "enrichGO, Molecular Function, bicluster 1")
goplot(ego_MF[[2]]) + labs(title = "enrichGO, Molecular Function, bicluster 2")
goplot(ego_MF[[3]]) + labs(title = "enrichGO, Molecular Function, bicluster 3")
goplot(ego_MF[[4]]) + labs(title = "enrichGO, Molecular Function, bicluster 4")
goplot(ego_MF[[5]]) + labs(title = "enrichGO, Molecular Function, bicluster 5")
goplot(ego_MF[[6]]) + labs(title = "enrichGO, Molecular Function, bicluster 8")
goplot(ego_MF[[7]]) + labs(title = "enrichGO, Molecular Function, bicluster 10")
goplot(ego_MF[[8]]) + labs(title = "enrichGO, Molecular Function, bicluster 12")
goplot(ego_MF[[9]]) + labs(title = "enrichGO, Molecular Function, bicluster 16")
goplot(ego_MF[[10]]) + labs(title = "enrichGO, Molecular Function, bicluster 18")
goplot(ego_MF[[11]]) + labs(title = "enrichGO, Molecular Function, bicluster 22")
goplot(ego_MF[[12]]) + labs(title = "enrichGO, Molecular Function, bicluster 26")
goplot(ego_MF[[13]]) + labs(title = "enrichGO, Molecular Function, bicluster 28")

goplot(ego_BP[[1]]) + labs(title = "enrichGO, Biological Process, bicluster 1")
goplot(ego_BP[[2]]) + labs(title = "enrichGO, Biological Process, bicluster 2")
goplot(ego_BP[[3]]) + labs(title = "enrichGO, Biological Process, bicluster 3")
goplot(ego_BP[[4]]) + labs(title = "enrichGO, Biological Process, bicluster 4")
goplot(ego_BP[[5]]) + labs(title = "enrichGO, Biological Process, bicluster 5")
goplot(ego_BP[[6]]) + labs(title = "enrichGO, Biological Process, bicluster 8")
goplot(ego_BP[[7]]) + labs(title = "enrichGO, Biological Process, bicluster 10")
goplot(ego_BP[[8]]) + labs(title = "enrichGO, Biological Process, bicluster 12")
goplot(ego_BP[[9]]) + labs(title = "enrichGO, Biological Process, bicluster 16")
goplot(ego_BP[[10]]) + labs(title = "enrichGO, Biological Process, bicluster 18")
goplot(ego_BP[[11]]) + labs(title = "enrichGO, Biological Process, bicluster 22")
goplot(ego_BP[[12]]) + labs(title = "enrichGO, Biological Process, bicluster 26")
goplot(ego_BP[[13]]) + labs(title = "enrichGO, Biological Process, bicluster 28")

goplot(ego_CC[[1]]) + labs(title = "enrichGO, Cellular Component, bicluster 1")
goplot(ego_CC[[2]]) + labs(title = "enrichGO, Cellular Component, bicluster 2")
goplot(ego_CC[[3]]) + labs(title = "enrichGO, Cellular Component, bicluster 3")
goplot(ego_CC[[4]]) + labs(title = "enrichGO, Cellular Component, bicluster 4")
goplot(ego_CC[[5]]) + labs(title = "enrichGO, Cellular Component, bicluster 5")
goplot(ego_CC[[6]]) + labs(title = "enrichGO, Cellular Component, bicluster 8")
goplot(ego_CC[[7]]) + labs(title = "enrichGO, Cellular Component, bicluster 10")
goplot(ego_CC[[8]]) + labs(title = "enrichGO, Cellular Component, bicluster 12")
goplot(ego_CC[[9]]) + labs(title = "enrichGO, Cellular Component, bicluster 16")
goplot(ego_CC[[10]]) + labs(title = "enrichGO, Cellular Component, bicluster 18")
goplot(ego_CC[[11]]) + labs(title = "enrichGO, Cellular Component, bicluster 22")
goplot(ego_CC[[12]]) + labs(title = "enrichGO, Cellular Component, bicluster 26")
goplot(ego_CC[[13]]) + labs(title = "enrichGO, Cellular Component, bicluster 28")

```


# 3. Export Pathway Analysis result to data

```{r}
pathway = list("ggo_MF" = ggo_MF,
               "ggo_BP" = ggo_BP,
               "ggo_CC" = ggo_CC,
               "ego_MF" = ego_MF,
               "ego_BP" = ego_BP,
               "ego_CC" = ego_CC)

saveRDS(pathway, file = "data/out_pathway/pathways_superbiclust.fabia.rds")
```
