==> 05k_Pathway_Analysis_GGM.Rmd <==
---
title: "Pathway Analysis for Gaussian Graphical Networks"
author: "John Zhuang"
date: "2024-06-17"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(multiMiR)
library(clusterProfiler)
library(org.Hs.eg.db)
library(stringr)
library(ggplot2)
```

# 0. Loading of Results

## A. List of all Gene names in the dataset

```{r}
gene.names = readRDS("data/derived/unique_gene_names.rds")
```

## B. Selected Features

```{r}
results.ggm= readRDS("data/sel_features/results_ggm_network_stats_all.rds")
```

```{r}
ggm.RPPA = results.ggm$full_sample$`RPPA Array`$`FDR1e-1`
names.ggm.RPPA = rownames(ggm.RPPA[which(ggm.RPPA$degree > 0),])

ggm.R2Gn.min = results.ggm$full_sample$RNASeq2GeneNorm$`FDR1e-14`
names.ggm.R2Gn.min = rownames(ggm.R2Gn.min[which(ggm.R2Gn.min$degree > mean(ggm.R2Gn.min$degree)),])
ggm.R2Gn.6 = results.ggm$full_sample$RNASeq2GeneNorm$`FDR1e-6`
names.ggm.R2Gn.6 = rownames(ggm.R2Gn.6[which(ggm.R2Gn.6$degree > mean(ggm.R2Gn.6$degree)),])

ggm.mRSG.min = results.ggm$full_sample$miRNASeqGene$`FDR1e-13`
names.ggm.mRSG.min = rownames(ggm.mRSG.min[which(ggm.mRSG.min$degree > mean(ggm.mRSG.min$degree)),])
ggm.mRSG.6 = results.ggm$full_sample$miRNASeqGene$`FDR1e-6`
names.ggm.mRSG.6 = rownames(ggm.mRSG.6[which(ggm.mRSG.6$degree > mean(ggm.mRSG.6$degree)),])

ggm.all.min = results.ggm$full_sample$all$`FDR1e-14`
names.ggm.all.min = rownames(ggm.all.min[which(ggm.all.min$degree > mean(ggm.all.min$degree)),])
ggm.all.6 = results.ggm$full_sample$all$`FDR1e-6`
names.ggm.all.6 = rownames(ggm.all.6[which(ggm.all.6$degree > mean(ggm.all.6$degree)),])

ggm.RPPA.a = results.ggm$by_vital_status$`RPPA Array`$alive$`FDR1e-1`
names.ggm.RPPA.a = rownames(ggm.RPPA.a[which(ggm.RPPA.a$degree > 0), ])

ggm.RPPA.d = results.ggm$by_vital_status$`RPPA Array`$dead$`FDR1e-1`
names.ggm.RPPA.d = rownames(ggm.RPPA.d[which(ggm.RPPA.d$degree > 0), ])

ggm.R2Gn.min.a = results.ggm$by_vital_status$RNASeq2GeneNorm$alive$`FDR1e-14`
names.ggm.R2Gn.min.a = rownames(ggm.R2Gn.min.a[which(ggm.R2Gn.min.a$degree > mean(ggm.R2Gn.min.a$degree)),])

ggm.R2Gn.min.d = results.ggm$by_vital_status$RNASeq2GeneNorm$dead$`FDR1e-14`
names.ggm.R2Gn.min.d = rownames(ggm.R2Gn.min.d[which(ggm.R2Gn.min.d$degree > mean(ggm.R2Gn.min.d$degree)),])

ggm.R2Gn.6.a = results.ggm$by_vital_status$RNASeq2GeneNorm$alive$`FDR1e-6`
names.ggm.R2Gn.6.a = rownames(ggm.R2Gn.6.a[which(ggm.R2Gn.6.a$degree > mean(ggm.R2Gn.6.a$degree)), ])

ggm.R2Gn.6.d = results.ggm$by_vital_status$RNASeq2GeneNorm$dead$`FDR1e-6`
names.ggm.R2Gn.6.d = rownames(ggm.R2Gn.6.d[which(ggm.R2Gn.6.d$degree > mean(ggm.R2Gn.6.d$degree)), ])

#ggm.RR.min.a = results.ggm$by_vital_status$`RPPA + R2Gn`$alive$`FDR1e-14`
#names.ggm.RR.min.a = rownames(ggm.RR.min.a[which(ggm.RR.min.a$degree > mean(ggm.RR.min.a$degree)), ])

#ggm.RR.min.d = results.ggm$by_vital_status$`RPPA + R2Gn`$dead$`FDR1e-14`
#names.ggm.RR.min.d = rownames(ggm.RR.min.d[which(ggm.RR.min.d$degree > mean(ggm.RR.min.d$degree)), ])

#ggm.RR.6.a = results.ggm$by_vital_status$`RPPA + R2Gn`$alive$`FDR1e-6`
#names.ggm.RR.6.a = rownames(ggm.RR.6.a[which(ggm.RR.6.a$degree > mean(ggm.RR.6.a$degree)), ])

#ggm.RR.6.d = results.ggm$by_vital_status$`RPPA + R2Gn`$dead$`FDR1e-6`
#names.ggm.RR.6.d = rownames(ggm.RR.6.d[which(ggm.RR.6.d$degree > mean(ggm.RR.6.d$degree)), ])

ggm.mRSG.min.a = results.ggm$by_vital_status$miRNASeqGene$alive$`FDR1e-13`
names.ggm.mRSG.min.a = rownames(ggm.mRSG.min.a[which(ggm.mRSG.min.a$degree > mean(ggm.mRSG.min.a$degree)),])

ggm.mRSG.min.d = results.ggm$by_vital_status$miRNASeqGene$dead$`FDR1e-13`
names.ggm.mRSG.min.d = rownames(ggm.mRSG.min.d[which(ggm.mRSG.min.d$degree > mean(ggm.mRSG.min.d$degree)),])

ggm.mRSG.6.a = results.ggm$by_vital_status$miRNASeqGene$alive$`FDR1e-6`
names.ggm.mRSG.6.a = rownames(ggm.mRSG.6.a[which(ggm.mRSG.6.a$degree > mean(ggm.mRSG.6.a$degree)), ])

ggm.mRSG.6.d = results.ggm$by_vital_status$miRNASeqGene$dead$`FDR1e-6`
names.ggm.mRSG.6.d = rownames(ggm.mRSG.6.d[which(ggm.mRSG.6.d$degree > mean(ggm.mRSG.6.d$degree)), ])

ggm.all.min.a = results.ggm$by_vital_status$all$alive$`FDR1e-14`
names.ggm.all.min.a = rownames(ggm.all.min.a[which(ggm.all.min.a$degree > mean(ggm.all.min.a$degree)),])

ggm.all.min.d = results.ggm$by_vital_status$all$dead$`FDR1e-14`
names.ggm.all.min.d = rownames(ggm.all.min.d[which(ggm.all.min.d$degree > mean(ggm.all.min.d$degree)),])

ggm.all.6.a = results.ggm$by_vital_status$all$alive$`FDR1e-6`
names.ggm.all.6.a = rownames(ggm.all.6.a[which(ggm.all.6.a$degree > mean(ggm.all.6.a$degree)), ])

ggm.all.6.d = results.ggm$by_vital_status$all$dead$`FDR1e-6`
names.ggm.all.6.d = rownames(ggm.all.6.d[which(ggm.all.6.d$degree > mean(ggm.all.6.d$degree)), ])
```


# 1. Convert microRNA to genes

Also remove the ending ".R2Gn" and ".RPPA" from gene names.


```{r}
split.by.omic = function(variable.list) {
  out.mRSG = variable.list[startsWith(variable.list, "hsa")]
  out.R2Gn = variable.list[endsWith(variable.list, ".R2Gn")]
  out.RPPA = variable.list[endsWith(variable.list, ".RPPA")]
  out = list("miRNASeqGene" = out.mRSG,
             "RNASeq2GeneNorm" = out.R2Gn,
             "RPPAArray" = out.RPPA)
  out
}
```

```{r}
multiomics.to.GENES = function(sel.features, full.search = F, table = "validated", drop.miRNA = F) {
  multiomics.list = split.by.omic(sel.features)
  # Convert miRNASeqGene to target GENE names
  # Skipped and returns an empty list instead of the list of select features,
  # or if microRNA features are explicitly dropped (drop.miRNA = T).
  # (Usually done when the features has only a few miRNA and unable to target any genes)
  if (length(multiomics.list[[1]]) > 0 & !drop.miRNA) {
    if (!full.search) { # Search only among the 198 genes in ACC, recommended
    targets= get_multimir(mirna = multiomics.list[[1]], 
                          target = gene.names,
                          table = table,
                          summary = T)
    mRSG.target.genes = unique(targets@data$target_symbol)
    } else { # Search the full dataset, not recommended. 
             # Use only if ACC gene search returns nothing or fails.
             # Genes not in the ACC genes need to be manually removed.
    targets= get_multimir(mirna = multiomics.list[[1]], 
                          table = table,
                          summary = T)  
    mRSG.target.genes = unique(targets@data$target_symbol)
    }
    
  } else {
    mRSG.target.genes = list()
  }
  #Remove suffix for Pathway Analysis Compatibility
  R2Gn.genes = str_remove(multiomics.list[[2]], ".R2Gn")
  
  RPPA.genes = str_remove(multiomics.list[[3]], ".RPPA")
  
  #Join all cleaned-up genes are remove duplicates.
  out1 = c(mRSG.target.genes, R2Gn.genes, RPPA.genes)
  out = as.character(unique(out1))
  out
}
```

```{r}
geneNames <- list("RPPA"= multiomics.to.GENES(names.ggm.RPPA),
                  "R2Gn.min"= multiomics.to.GENES(names.ggm.R2Gn.min),
                  "R2Gn.6"= multiomics.to.GENES(names.ggm.R2Gn.6),
                  "mRSG.min"= multiomics.to.GENES(names.ggm.mRSG.min),
                  "mRSG.6"= multiomics.to.GENES(names.ggm.mRSG.6),
                  "all.min"= multiomics.to.GENES(names.ggm.all.min),
                  "all.6"= multiomics.to.GENES(names.ggm.all.6),
                  
                  "RPPA.0" = multiomics.to.GENES(names.ggm.RPPA.a),
                  "RPPA.1" = multiomics.to.GENES(names.ggm.RPPA.d),
                  "R2Gn.min.0" = multiomics.to.GENES(names.ggm.R2Gn.min.a),
                  "R2Gn.min.1" = multiomics.to.GENES(names.ggm.R2Gn.min.d),
                  "R2Gn.6.0" = multiomics.to.GENES(names.ggm.R2Gn.6.a),
                  "R2Gn.6.1" = multiomics.to.GENES(names.ggm.R2Gn.6.d),
                  #"RR.min.0" = multiomics.to.GENES(names.ggm.RR.min.a),
                  #"RR.min.1" = multiomics.to.GENES(names.ggm.RR.min.d),
                  #"RR.6.0" = multiomics.to.GENES(names.ggm.RR.6.a),
                  #"RR.6.1" = multiomics.to.GENES(names.ggm.RR.6.d)
                  "mRSG.min.0" = multiomics.to.GENES(names.ggm.mRSG.min.a),
                  "mRSG.min.1" = multiomics.to.GENES(names.ggm.mRSG.min.d),
                  "mRSG.6.0" = multiomics.to.GENES(names.ggm.mRSG.6.a),
                  "mRSG.6.1" = multiomics.to.GENES(names.ggm.mRSG.6.d),
                  "all.min.0" = multiomics.to.GENES(names.ggm.all.min.a),
                  "all.min.1" = multiomics.to.GENES(names.ggm.all.min.d),
                  "all.6.0" = multiomics.to.GENES(names.ggm.all.6.a),
                  "all.6.1" = multiomics.to.GENES(names.ggm.all.6.d)
                  )
```
```{r}
geneNames
```


```{r}
saveRDS(geneNames, file = "data/sel_features/gene_converted/genes_ggm.rds")
```



# 2. Pathway Analysis

## A. GO Classification

```{r}
ggo_MF <- list()#functional profile of a geneset for Molecular Function
ggo_BP <- list()#functional profile of a geneset for Biological Process
ggo_CC <- list()#functional profile of a geneset for Cellular Component


for(z in 1:length(geneNames)){
  print(z)
  input <- geneNames[[z]]

  ggo_MF[[z]] <- groupGO(gene=input,
    OrgDb = org.Hs.eg.db,
    ont="MF",
    keyType="SYMBOL",
    readable=TRUE)
  
  ggo_BP[[z]] <- groupGO(gene=input,
    OrgDb = org.Hs.eg.db,
    ont="BP",
    keyType="SYMBOL",
    readable=TRUE)

  ggo_CC[[z]] <- groupGO(gene=input,
    OrgDb = org.Hs.eg.db,
    ont="CC",
    keyType="SYMBOL",
    readable=TRUE)    
}
```

```{r}
for(z in 1:length(geneNames)) {
  print(head(ggo_MF[[z]]))
  print(head(ggo_BP[[z]]))
  print(head(ggo_CC[[z]]))
}
```
```{r}
barplot(ggo_MF[[1]], title = "groupGO, Molecular Function, \n RPPAArray, FDR = 0.9")
barplot(ggo_MF[[2]], title = "groupGO, Molecular Function, \n RNASeq2GeneNorm, FDR = 1-1e-14")
barplot(ggo_MF[[3]], title = "groupGO, Molecular Function, \n RNASeq2GeneNorm, FDR = 1-1e-6")
barplot(ggo_MF[[4]], title = "groupGO, Molecular Function, \n miRNASeqGene, FDR = 1-1e-13")
barplot(ggo_MF[[5]], title = "groupGO, Molecular Function, \n miRNASeqGene, FDR = 1-1e-6")
barplot(ggo_MF[[6]], title = "groupGO, Molecular Function, \n All datasets, FDR = 1-1e-14")
barplot(ggo_MF[[7]], title = "groupGO, Molecular Function, \n All datasets, FDR = 1-1e-6")
barplot(ggo_MF[[8]], title = "groupGO, Molecular Function, \n RPPAArray, Surviving Patients, FDR = 0.9")
barplot(ggo_MF[[9]], title = "groupGO, Molecular Function, \n RPPAArray, Deceased Patients, FDR = 0.9")
barplot(ggo_MF[[10]], title = "groupGO, Molecular Function, \n RNASeq2GeneNorm, Surviving Patients, FDR = 1-1e-14")
barplot(ggo_MF[[11]], title = "groupGO, Molecular Function, \n RNASeq2GeneNorm, Deceased Patients, FDR = 1-1e-14")
barplot(ggo_MF[[12]], title = "groupGO, Molecular Function, \n RNASeq2GeneNorm, Surviving Patients, FDR = 1-1e-6")
barplot(ggo_MF[[13]], title = "groupGO, Molecular Function, \n RNASeq2GeneNorm, Deceased Patients, FDR = 1-1e-6")
barplot(ggo_MF[[14]], title = "groupGO, Molecular Function, \n miRNASeqGene, Surviving Patients, FDR = 1-1e-13")
barplot(ggo_MF[[15]], title = "groupGO, Molecular Function, \n miRNASeqGene, Deceased Patients, FDR = 1-1e-13")
barplot(ggo_MF[[16]], title = "groupGO, Molecular Function, \n miRNASeqGene, Surviving Patients, FDR = 1-1e-6")
barplot(ggo_MF[[17]], title = "groupGO, Molecular Function, \n miRNASeqGene, Deceased Patients, FDR = 1-1e-6")
barplot(ggo_MF[[18]], title = "groupGO, Molecular Function, \n All datasets, Surviving Patients, FDR = 1-1e-14")
barplot(ggo_MF[[19]], title = "groupGO, Molecular Function, \n All datasets, Deceased Patients, FDR = 1-1e-14")
barplot(ggo_MF[[20]], title = "groupGO, Molecular Function, \n All datasets, Surviving Patients, FDR = 1-1e-6")
barplot(ggo_MF[[21]], title = "groupGO, Molecular Function, \n All datasets, Deceased Patients, FDR = 1-1e-6")

barplot(ggo_BP[[1]], title = "groupGO, Biological Process, \n RPPAArray, FDR = 0.9")
barplot(ggo_BP[[2]], title = "groupGO, Biological Process, \n RNASeq2GeneNorm, FDR = 1-1e-14")
barplot(ggo_BP[[3]], title = "groupGO, Biological Process, \n RNASeq2GeneNorm, FDR = 1-1e-6")
barplot(ggo_BP[[4]], title = "groupGO, Biological Process, \n miRNASeqGene, FDR = 1-1e-13")
barplot(ggo_BP[[5]], title = "groupGO, Biological Process, \n miRNASeqGene, FDR = 1-1e-6")
barplot(ggo_BP[[6]], title = "groupGO, Biological Process, \n All datasets, FDR = 1-1e-14")
barplot(ggo_BP[[7]], title = "groupGO, Biological Process, \n All datasets, FDR = 1-1e-6")
barplot(ggo_BP[[8]], title = "groupGO, Biological Process, \n RPPAArray, Surviving Patients, FDR = 0.9")
barplot(ggo_BP[[9]], title = "groupGO, Biological Process, \n RPPAArray, Deceased Patients, FDR = 0.9")
barplot(ggo_BP[[10]], title = "groupGO, Biological Process, \n RNASeq2GeneNorm, Surviving Patients, FDR = 1-1e-14")
barplot(ggo_BP[[11]], title = "groupGO, Biological Process, \n RNASeq2GeneNorm, Deceased Patients, FDR = 1-1e-14")
barplot(ggo_BP[[12]], title = "groupGO, Biological Process, \n RNASeq2GeneNorm, Surviving Patients, FDR = 1-1e-6")
barplot(ggo_BP[[13]], title = "groupGO, Biological Process, \n RNASeq2GeneNorm, Deceased Patients, FDR = 1-1e-6")
barplot(ggo_BP[[14]], title = "groupGO, Biological Process, \n miRNASeqGene, Surviving Patients, FDR = 1-1e-14")
barplot(ggo_BP[[15]], title = "groupGO, Biological Process, \n miRNASeqGene, Deceased Patients, FDR = 1-1e-14")
barplot(ggo_BP[[16]], title = "groupGO, Biological Process, \n miRNASeqGene, Surviving Patients, FDR = 1-1e-6")
barplot(ggo_BP[[17]], title = "groupGO, Biological Process, \n miRNASeqGene, Deceased Patients, FDR = 1-1e-6")
barplot(ggo_BP[[18]], title = "groupGO, Biological Process, \n All datasets, Surviving Patients, FDR = 1-1e-14")
barplot(ggo_BP[[19]], title = "groupGO, Biological Process, \n All datasets, Deceased Patients, FDR = 1-1e-14")
barplot(ggo_BP[[20]], title = "groupGO, Biological Process, \n All datasets, Surviving Patients, FDR = 1-1e-6")
barplot(ggo_BP[[21]], title = "groupGO, Biological Process, \n All datasets, Deceased Patients, FDR = 1-1e-6")

barplot(ggo_CC[[1]], title = "groupGO, Cellular Component, \n RPPAArray, FDR = 0.9")
barplot(ggo_CC[[2]], title = "groupGO, Cellular Component, \n RNASeq2GeneNorm, FDR = 1-1e-14")
barplot(ggo_CC[[3]], title = "groupGO, Cellular Component, \n RNASeq2GeneNorm, FDR = 1-1e-6")
barplot(ggo_CC[[4]], title = "groupGO, Cellular Component, \n miRNASeqGene, FDR = 1-1e-13")
barplot(ggo_CC[[5]], title = "groupGO, Cellular Component, \n miRNASeqGene, FDR = 1-1e-6")
barplot(ggo_CC[[6]], title = "groupGO, Cellular Component, \n All datasets, FDR = 1-1e-14")
barplot(ggo_CC[[7]], title = "groupGO, Cellular Component, \n All datasets, FDR = 1-1e-6")
barplot(ggo_CC[[8]], title = "groupGO, Cellular Component, \n RPPAArray, Surviving Patients, FDR = 0.9")
barplot(ggo_CC[[9]], title = "groupGO, Cellular Component, \n RPPAArray, Deceased Patients, FDR = 0.9")
barplot(ggo_CC[[10]], title = "groupGO, Cellular Component, \n RNASeq2GeneNorm, Surviving Patients, FDR = 1-1e-14")
barplot(ggo_CC[[11]], title = "groupGO, Cellular Component, \n RNASeq2GeneNorm, Deceased Patients, FDR = 1-1e-14")
barplot(ggo_CC[[12]], title = "groupGO, Cellular Component, \n RNASeq2GeneNorm, Surviving Patients, FDR = 1-1e-6")
barplot(ggo_CC[[13]], title = "groupGO, Cellular Component, \n RNASeq2GeneNorm, Deceased Patients, FDR = 1-1e-6")
barplot(ggo_CC[[14]], title = "groupGO, Cellular Component, \n miRNASeqGene, Surviving Patients, FDR = 1-1e-13")
barplot(ggo_CC[[15]], title = "groupGO, Cellular Component, \n miRNASeqGene, Deceased Patients, FDR = 1-1e-13")
barplot(ggo_CC[[16]], title = "groupGO, Cellular Component, \n miRNASeqGene, Surviving Patients, FDR = 1-1e-6")
barplot(ggo_CC[[17]], title = "groupGO, Cellular Component, \n miRNASeqGene, Deceased Patients, FDR = 1-1e-6")
barplot(ggo_CC[[18]], title = "groupGO, Cellular Component, \n All datasets, Surviving Patients, FDR = 1-1e-14")
barplot(ggo_CC[[19]], title = "groupGO, Cellular Component, \n All datasets, Deceased Patients, FDR = 1-1e-14")
barplot(ggo_CC[[20]], title = "groupGO, Cellular Component, \n All datasets, Surviving Patients, FDR = 1-1e-6")
barplot(ggo_CC[[21]], title = "groupGO, Cellular Component, \n All datasets, Deceased Patients, FDR = 1-1e-6")


```

## B. GO Over-Representation

```{r}
ego_MF <- list()#GO enrichment of a geneset for Molecular Function
ego_BP <- list()#GO enrichment of a geneset for Biological Process
ego_CC <- list()#GO enrichment of a geneset for Cellular Component


for(z in 1:length(geneNames)){
  print(z)
  input <- geneNames[[z]]
  ego_MF[[z]] <- enrichGO(
    gene          = input,
    OrgDb         = org.Hs.eg.db,
    ont           = "MF",
    pAdjustMethod = "BH",
    keyType="SYMBOL",
    pvalueCutoff  = 0.05,
    qvalueCutoff  = 0.05,
    minGSSize     = 2,
    readable      = TRUE)
  
  ego_BP[[z]] <- enrichGO(
    gene          = input,
    OrgDb         = org.Hs.eg.db,
    ont           = "BP",
    pAdjustMethod = "BH",
    keyType="SYMBOL",
    pvalueCutoff  = 0.05,
    qvalueCutoff  = 0.05,
    minGSSize     = 2,
    readable      = TRUE)
  
  ego_CC[[z]] <- enrichGO(
    gene          = input,
    OrgDb         = org.Hs.eg.db,
    ont           = "CC",
    pAdjustMethod = "BH",
    keyType="SYMBOL",
    pvalueCutoff  = 0.05,
    qvalueCutoff  = 0.05,
    minGSSize     = 2,
    readable      = TRUE)
}
```

```{r}
for(z in 1:length(geneNames)){
  print(head(ego_MF[[z]]))
  print(head(ego_BP[[z]]))
  print(head(ego_CC[[z]]))
}
```


```{r}
barplot(ego_MF[[1]], title = "groupGO, Molecular Function, \n RPPAArray, FDR = 0.9")
barplot(ego_MF[[2]], title = "groupGO, Molecular Function, \n RNASeq2GeneNorm, FDR = 1-1e-14")
barplot(ego_MF[[3]], title = "groupGO, Molecular Function, \n RNASeq2GeneNorm, FDR = 1-1e-6")
barplot(ego_MF[[4]], title = "groupGO, Molecular Function, \n miRNASeqGene, FDR = 1-1e-13")
barplot(ego_MF[[5]], title = "groupGO, Molecular Function, \n miRNASeqGene, FDR = 1-1e-6")
barplot(ego_MF[[6]], title = "groupGO, Molecular Function, \n All datasets, FDR = 1-1e-14")
barplot(ego_MF[[7]], title = "groupGO, Molecular Function, \n All datasets, FDR = 1-1e-6")
barplot(ego_MF[[8]], title = "groupGO, Molecular Function, \n RPPAArray, Surviving Patients, FDR = 0.9")
barplot(ego_MF[[9]], title = "groupGO, Molecular Function, \n RPPAArray, Deceased Patients, FDR = 0.9")
barplot(ego_MF[[10]], title = "groupGO, Molecular Function, \n RNASeq2GeneNorm, Surviving Patients, FDR = 1-1e-14")
barplot(ego_MF[[11]], title = "groupGO, Molecular Function, \n RNASeq2GeneNorm, Deceased Patients, FDR = 1-1e-14")
barplot(ego_MF[[12]], title = "groupGO, Molecular Function, \n RNASeq2GeneNorm, Surviving Patients, FDR = 1-1e-6")
barplot(ego_MF[[13]], title = "groupGO, Molecular Function, \n RNASeq2GeneNorm, Deceased Patients, FDR = 1-1e-6")
barplot(ego_MF[[14]], title = "groupGO, Molecular Function, \n miRNASeqGene, Surviving Patients, FDR = 1-1e-13")
barplot(ego_MF[[15]], title = "groupGO, Molecular Function, \n miRNASeqGene, Deceased Patients, FDR = 1-1e-13")
barplot(ego_MF[[16]], title = "groupGO, Molecular Function, \n miRNASeqGene, Surviving Patients, FDR = 1-1e-6")
barplot(ego_MF[[17]], title = "groupGO, Molecular Function, \n miRNASeqGene, Deceased Patients, FDR = 1-1e-6")
barplot(ego_MF[[18]], title = "groupGO, Molecular Function, \n All datasets, Surviving Patients, FDR = 1-1e-14")
barplot(ego_MF[[19]], title = "groupGO, Molecular Function, \n All datasets, Deceased Patients, FDR = 1-1e-14")
barplot(ego_MF[[20]], title = "groupGO, Molecular Function, \n All datasets, Surviving Patients, FDR = 1-1e-6")
barplot(ego_MF[[21]], title = "groupGO, Molecular Function, \n All datasets, Deceased Patients, FDR = 1-1e-6")

barplot(ego_BP[[1]], title = "groupGO, Biological Process, \n RPPAArray, FDR = 0.9")
barplot(ego_BP[[2]], title = "groupGO, Biological Process, \n RNASeq2GeneNorm, FDR = 1-1e-14")
barplot(ego_BP[[3]], title = "groupGO, Biological Process, \n RNASeq2GeneNorm, FDR = 1-1e-6")
barplot(ego_BP[[4]], title = "groupGO, Biological Process, \n miRNASeqGene, FDR = 1-1e-13")
barplot(ego_BP[[5]], title = "groupGO, Biological Process, \n miRNASeqGene, FDR = 1-1e-6")
barplot(ego_BP[[6]], title = "groupGO, Biological Process, \n All datasets, FDR = 1-1e-14")
barplot(ego_BP[[7]], title = "groupGO, Biological Process, \n All datasets, FDR = 1-1e-6")
barplot(ego_BP[[8]], title = "groupGO, Biological Process, \n RPPAArray, Surviving Patients, FDR = 0.9")
barplot(ego_BP[[9]], title = "groupGO, Biological Process, \n RPPAArray, Deceased Patients, FDR = 0.9")
barplot(ego_BP[[10]], title = "groupGO, Biological Process, \n RNASeq2GeneNorm, Surviving Patients, FDR = 1-1e-14")
barplot(ego_BP[[11]], title = "groupGO, Biological Process, \n RNASeq2GeneNorm, Deceased Patients, FDR = 1-1e-14")
barplot(ego_BP[[12]], title = "groupGO, Biological Process, \n RNASeq2GeneNorm, Surviving Patients, FDR = 1-1e-6")
barplot(ego_BP[[13]], title = "groupGO, Biological Process, \n RNASeq2GeneNorm, Deceased Patients, FDR = 1-1e-6")
barplot(ego_BP[[14]], title = "groupGO, Biological Process, \n miRNASeqGene, Surviving Patients, FDR = 1-1e-14")
barplot(ego_BP[[15]], title = "groupGO, Biological Process, \n miRNASeqGene, Deceased Patients, FDR = 1-1e-14")
barplot(ego_BP[[16]], title = "groupGO, Biological Process, \n miRNASeqGene, Surviving Patients, FDR = 1-1e-6")
barplot(ego_BP[[17]], title = "groupGO, Biological Process, \n miRNASeqGene, Deceased Patients, FDR = 1-1e-6")
barplot(ego_BP[[18]], title = "groupGO, Biological Process, \n All datasets, Surviving Patients, FDR = 1-1e-14")
barplot(ego_BP[[19]], title = "groupGO, Biological Process, \n All datasets, Deceased Patients, FDR = 1-1e-14")
barplot(ego_BP[[20]], title = "groupGO, Biological Process, \n All datasets, Surviving Patients, FDR = 1-1e-6")
barplot(ego_BP[[21]], title = "groupGO, Biological Process, \n All datasets, Deceased Patients, FDR = 1-1e-6")

barplot(ego_CC[[1]], title = "groupGO, Cellular Component, \n RPPAArray, FDR = 0.9")
barplot(ego_CC[[2]], title = "groupGO, Cellular Component, \n RNASeq2GeneNorm, FDR = 1-1e-14")
barplot(ego_CC[[3]], title = "groupGO, Cellular Component, \n RNASeq2GeneNorm, FDR = 1-1e-6")
barplot(ego_CC[[4]], title = "groupGO, Cellular Component, \n miRNASeqGene, FDR = 1-1e-13")
barplot(ego_CC[[5]], title = "groupGO, Cellular Component, \n miRNASeqGene, FDR = 1-1e-6")
barplot(ego_CC[[6]], title = "groupGO, Cellular Component, \n All datasets, FDR = 1-1e-14")
barplot(ego_CC[[7]], title = "groupGO, Cellular Component, \n All datasets, FDR = 1-1e-6")
barplot(ego_CC[[8]], title = "groupGO, Cellular Component, \n RPPAArray, Surviving Patients, FDR = 0.9")
barplot(ego_CC[[9]], title = "groupGO, Cellular Component, \n RPPAArray, Deceased Patients, FDR = 0.9")
barplot(ego_CC[[10]], title = "groupGO, Cellular Component, \n RNASeq2GeneNorm, Surviving Patients, FDR = 1-1e-14")
barplot(ego_CC[[11]], title = "groupGO, Cellular Component, \n RNASeq2GeneNorm, Deceased Patients, FDR = 1-1e-14")
barplot(ego_CC[[12]], title = "groupGO, Cellular Component, \n RNASeq2GeneNorm, Surviving Patients, FDR = 1-1e-6")
barplot(ego_CC[[13]], title = "groupGO, Cellular Component, \n RNASeq2GeneNorm, Deceased Patients, FDR = 1-1e-6")
barplot(ego_CC[[14]], title = "groupGO, Cellular Component, \n miRNASeqGene, Surviving Patients, FDR = 1-1e-13")
barplot(ego_CC[[15]], title = "groupGO, Cellular Component, \n miRNASeqGene, Deceased Patients, FDR = 1-1e-13")
barplot(ego_CC[[16]], title = "groupGO, Cellular Component, \n miRNASeqGene, Surviving Patients, FDR = 1-1e-6")
barplot(ego_CC[[17]], title = "groupGO, Cellular Component, \n miRNASeqGene, Deceased Patients, FDR = 1-1e-6")
barplot(ego_CC[[18]], title = "groupGO, Cellular Component, \n All datasets, Surviving Patients, FDR = 1-1e-14")
barplot(ego_CC[[19]], title = "groupGO, Cellular Component, \n All datasets, Deceased Patients, FDR = 1-1e-14")
barplot(ego_CC[[20]], title = "groupGO, Cellular Component, \n All datasets, Surviving Patients, FDR = 1-1e-6")
barplot(ego_CC[[21]], title = "groupGO, Cellular Component, \n All datasets, Deceased Patients, FDR = 1-1e-6")




```

### DAG plots for ego

```{r fig.width=20, fig.height=10}
goplot(ego_MF[[1]]) + labs(title = "groupGO, Molecular Function, \n RPPAArray, FDR = 0.9")
goplot(ego_MF[[2]]) + labs(title = "groupGO, Molecular Function, \n RNASeq2GeneNorm, FDR = 1-1e-14")
goplot(ego_MF[[3]]) + labs(title = "groupGO, Molecular Function, \n RNASeq2GeneNorm, FDR = 1-1e-6")
goplot(ego_MF[[4]]) + labs(title = "groupGO, Molecular Function, \n miRNASeqGene, FDR = 1-1e-13")
goplot(ego_MF[[5]]) + labs(title = "groupGO, Molecular Function, \n miRNASeqGene, FDR = 1-1e-6")
goplot(ego_MF[[6]]) + labs(title = "groupGO, Molecular Function, \n All datasets, FDR = 1-1e-14")
goplot(ego_MF[[7]]) + labs(title = "groupGO, Molecular Function, \n All datasets, FDR = 1-1e-6")
goplot(ego_MF[[8]]) + labs(title = "groupGO, Molecular Function, \n RPPAArray, Surviving Patients, FDR = 0.9")
goplot(ego_MF[[9]]) + labs(title = "groupGO, Molecular Function, \n RPPAArray, Deceased Patients, FDR = 0.9")
goplot(ego_MF[[10]]) + labs(title = "groupGO, Molecular Function, \n RNASeq2GeneNorm, Surviving Patients, FDR = 1-1e-14")
goplot(ego_MF[[11]]) + labs(title = "groupGO, Molecular Function, \n RNASeq2GeneNorm, Deceased Patients, FDR = 1-1e-14")
goplot(ego_MF[[12]]) + labs(title = "groupGO, Molecular Function, \n RNASeq2GeneNorm, Surviving Patients, FDR = 1-1e-6")
goplot(ego_MF[[13]]) + labs(title = "groupGO, Molecular Function, \n RNASeq2GeneNorm, Deceased Patients, FDR = 1-1e-6")
goplot(ego_MF[[14]]) + labs(title = "groupGO, Molecular Function, \n miRNASeqGene, Surviving Patients, FDR = 1-1e-13")
goplot(ego_MF[[15]]) + labs(title = "groupGO, Molecular Function, \n miRNASeqGene, Deceased Patients, FDR = 1-1e-13")
goplot(ego_MF[[16]]) + labs(title = "groupGO, Molecular Function, \n miRNASeqGene, Surviving Patients, FDR = 1-1e-6")
goplot(ego_MF[[17]]) + labs(title = "groupGO, Molecular Function, \n miRNASeqGene, Deceased Patients, FDR = 1-1e-6")
goplot(ego_MF[[18]]) + labs(title = "groupGO, Molecular Function, \n All datasets, Surviving Patients, FDR = 1-1e-14")
goplot(ego_MF[[19]]) + labs(title = "groupGO, Molecular Function, \n All datasets, Deceased Patients, FDR = 1-1e-14")
goplot(ego_MF[[20]]) + labs(title = "groupGO, Molecular Function, \n All datasets, Surviving Patients, FDR = 1-1e-6")
goplot(ego_MF[[21]]) + labs(title = "groupGO, Molecular Function, \n All datasets, Deceased Patients, FDR = 1-1e-6")

goplot(ego_BP[[1]]) + labs(title = "groupGO, Biological Process, \n RPPAArray, FDR = 0.9")
goplot(ego_BP[[2]]) + labs(title = "groupGO, Biological Process, \n RNASeq2GeneNorm, FDR = 1-1e-14")
goplot(ego_BP[[3]]) + labs(title = "groupGO, Biological Process, \n RNASeq2GeneNorm, FDR = 1-1e-6")
goplot(ego_BP[[4]]) + labs(title = "groupGO, Biological Process, \n miRNASeqGene, FDR = 1-1e-13")
goplot(ego_BP[[5]]) + labs(title = "groupGO, Biological Process, \n miRNASeqGene, FDR = 1-1e-6")
goplot(ego_BP[[6]]) + labs(title = "groupGO, Biological Process, \n All datasets, FDR = 1-1e-14")
goplot(ego_BP[[7]]) + labs(title = "groupGO, Biological Process, \n All datasets, FDR = 1-1e-6")
goplot(ego_BP[[8]]) + labs(title = "groupGO, Biological Process, \n RPPAArray, Surviving Patients, FDR = 0.9")
goplot(ego_BP[[9]]) + labs(title = "groupGO, Biological Process, \n RPPAArray, Deceased Patients, FDR = 0.9")
goplot(ego_BP[[10]]) + labs(title = "groupGO, Biological Process, \n RNASeq2GeneNorm, Surviving Patients, FDR = 1-1e-14")
goplot(ego_BP[[11]]) + labs(title = "groupGO, Biological Process, \n RNASeq2GeneNorm, Deceased Patients, FDR = 1-1e-14")
goplot(ego_BP[[12]]) + labs(title = "groupGO, Biological Process, \n RNASeq2GeneNorm, Surviving Patients, FDR = 1-1e-6")
goplot(ego_BP[[13]]) + labs(title = "groupGO, Biological Process, \n RNASeq2GeneNorm, Deceased Patients, FDR = 1-1e-6")
goplot(ego_BP[[14]]) + labs(title = "groupGO, Biological Process, \n miRNASeqGene, Surviving Patients, FDR = 1-1e-14")
goplot(ego_BP[[15]]) + labs(title = "groupGO, Biological Process, \n miRNASeqGene, Deceased Patients, FDR = 1-1e-14")
goplot(ego_BP[[16]]) + labs(title = "groupGO, Biological Process, \n miRNASeqGene, Surviving Patients, FDR = 1-1e-6")
goplot(ego_BP[[17]]) + labs(title = "groupGO, Biological Process, \n miRNASeqGene, Deceased Patients, FDR = 1-1e-6")
goplot(ego_BP[[18]]) + labs(title = "groupGO, Biological Process, \n All datasets, Surviving Patients, FDR = 1-1e-14")
goplot(ego_BP[[19]]) + labs(title = "groupGO, Biological Process, \n All datasets, Deceased Patients, FDR = 1-1e-14")
goplot(ego_BP[[20]]) + labs(title = "groupGO, Biological Process, \n All datasets, Surviving Patients, FDR = 1-1e-6")
goplot(ego_BP[[21]]) + labs(title = "groupGO, Biological Process, \n All datasets, Deceased Patients, FDR = 1-1e-6")

goplot(ego_CC[[1]]) + labs(title = "groupGO, Cellular Component, \n RPPAArray, FDR = 0.9")
goplot(ego_CC[[2]]) + labs(title = "groupGO, Cellular Component, \n RNASeq2GeneNorm, FDR = 1-1e-14")
goplot(ego_CC[[3]]) + labs(title = "groupGO, Cellular Component, \n RNASeq2GeneNorm, FDR = 1-1e-6")
goplot(ego_CC[[4]]) + labs(title = "groupGO, Cellular Component, \n miRNASeqGene, FDR = 1-1e-13")
goplot(ego_CC[[5]]) + labs(title = "groupGO, Cellular Component, \n miRNASeqGene, FDR = 1-1e-6")
goplot(ego_CC[[6]]) + labs(title = "groupGO, Cellular Component, \n All datasets, FDR = 1-1e-14")
goplot(ego_CC[[7]]) + labs(title = "groupGO, Cellular Component, \n All datasets, FDR = 1-1e-6")
goplot(ego_CC[[8]]) + labs(title = "groupGO, Cellular Component, \n RPPAArray, Surviving Patients, FDR = 0.9")
goplot(ego_CC[[9]]) + labs(title = "groupGO, Cellular Component, \n RPPAArray, Deceased Patients, FDR = 0.9")
goplot(ego_CC[[10]]) + labs(title = "groupGO, Cellular Component, \n RNASeq2GeneNorm, Surviving Patients, FDR = 1-1e-14")
goplot(ego_CC[[11]]) + labs(title = "groupGO, Cellular Component, \n RNASeq2GeneNorm, Deceased Patients, FDR = 1-1e-14")
goplot(ego_CC[[12]]) + labs(title = "groupGO, Cellular Component, \n RNASeq2GeneNorm, Surviving Patients, FDR = 1-1e-6")
goplot(ego_CC[[13]]) + labs(title = "groupGO, Cellular Component, \n RNASeq2GeneNorm, Deceased Patients, FDR = 1-1e-6")
goplot(ego_CC[[14]]) + labs(title = "groupGO, Cellular Component, \n miRNASeqGene, Surviving Patients, FDR = 1-1e-13")
goplot(ego_CC[[15]]) + labs(title = "groupGO, Cellular Component, \n miRNASeqGene, Deceased Patients, FDR = 1-1e-13")
goplot(ego_CC[[16]]) + labs(title = "groupGO, Cellular Component, \n miRNASeqGene, Surviving Patients, FDR = 1-1e-6")
goplot(ego_CC[[17]]) + labs(title = "groupGO, Cellular Component, \n miRNASeqGene, Deceased Patients, FDR = 1-1e-6")
goplot(ego_CC[[18]]) + labs(title = "groupGO, Cellular Component, \n All datasets, Surviving Patients, FDR = 1-1e-14")
goplot(ego_CC[[19]]) + labs(title = "groupGO, Cellular Component, \n All datasets, Deceased Patients, FDR = 1-1e-14")
goplot(ego_CC[[20]]) + labs(title = "groupGO, Cellular Component, \n All datasets, Surviving Patients, FDR = 1-1e-6")
goplot(ego_CC[[21]]) + labs(title = "groupGO, Cellular Component, \n All datasets, Deceased Patients, FDR = 1-1e-6")




```

# 3. Export Pathway Analysis result to data

```{r}
pathway = list("ggo_MF" = ggo_MF,
               "ggo_BP" = ggo_BP,
               "ggo_CC" = ggo_CC,
               "ego_MF" = ego_MF,
               "ego_BP" = ego_BP,
               "ego_CC" = ego_CC)

saveRDS(pathway, file = "data/out_pathway/pathways_ggm.rds")
```
